Refactoring auth.py
